requirements.txt
click==6.7
Flask==1.0.2
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
python-dateutil==2.7.3
pytz==2018.5
RPi.GPIO==0.6.3
six==1.11.0
Werkzeug==0.14.1
